Swap gas optimizations R1-R4: skip unused work on the hot path
Measured per-tx (cold, steady-state; snapshots/SwapGasProfile.json):
hooked bucket-moving swap 123.3k -> 91.1k; auction non-winner 141.7k -> 112.2k;
overheads vs a hookless pool now +15.6k (hooked, no triggers) and +36.7k
(auction) from +47.7k / +66.1k.

- R1: _afterSwap skips the oracle bound, list walks, and cursor write when the
  pool has no registered triggers (the dominant costs; the oracle bound only
  protects trigger execution). The cursor is deliberately left stale on the
  fast path; _addPositionTriggers re-baselines it to the current bucket when
  the first trigger registers, so stale-cursor replays can never fire a
  trigger spuriously (already-met conditions are handled by the pre-existing
  config-time immediate execution). Oracle try/catch factored into
  _tryOracleMaxEndTick; liquidity-sender gate deduped (EIP-170 headroom).
  Regression: testTriggerRegistrationRebaselinesStaleCursor.
- R2: _beforeSwap and _notifyAuctionLiquidityChange skip the auction controller
  for static-fee pools (auctions require DYNAMIC_FEE_FLAG) - a pure calldata
  check removing the controller round trip from every default-pool swap.
- R3: PoolAuctionConfig reordered so slot 0 carries the whole swap hot path
  (gate, epoch sync, drip throttle, winner fee, biddingEnabled); reserve
  packed to uint96; controller internals use storage pointers instead of
  3-slot memory copies.
- R4: same-TRANSACTION transient memo (EIP-1153 clears per tx, not per block):
  once a pool is synced+dripped at a timestamp, repeat touches - aggregator
  split routes, multi-hop, and the 2-3 controller pings of every automation
  action in a trigger cascade - skip straight to the winner check.

Gas ceilings tightened (60k/70k); auction warm pins now 42.3k/45.6k (from
61.6k/65.9k). Hook 24,527 B (+49); suite 555/555 (+1 pre-existing fork test).

+++ test/hook/invariants/HookAuctionControllerInvariants.t.sol
@@ -38,7 +38,7 @@ contract HookAuctionControllerHandler is BaseTest {
     using PoolIdLibrary for PoolKey;
 
     uint32 public constant EPOCH_LENGTH = 3600;
-    uint128 public constant RESERVE = 1e15;
+    uint96 public constant RESERVE = 1e15;
 
     HookAuctionController public auctionController;
     RevertHook public hook;
